refactor(player-and-printer): rename and move player and printer class
