+++ src/main/java/racingcar/Printer.java
@@ -1,6 +1,6 @@
-package racingcar.consol;
+package racingcar;
 
-public class OutputConsole {
+public class Printer {
     private static final String INPUT_CAR_NAME_MESSAGE = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
     private static final String INPUT_TRY_NUMBER_MESSAGE = "시도할 횟수는 몇회인가요?";
     private static final String MOVING_RESULT_HEADER_MESSAGE = "실행 결과";
